+++ config/medium/BeautifulBaseTheMostBeautiful/code/CGCode.ts
@@ -1,0 +1,6 @@
+const n: string = readline();
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('The most beautiful base of n.');