+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "key-value store".
 - configuration for "The Lord of the Annuli".
 - configuration for "Mondrian painting strokes".
+- configuration for "Beautiful base : The most beautiful".
 
 ## [1.16.0] - 2025-06-30
 ### Added